Move migrate function to Migrations file. Use request param for latestSupportedVersion
